requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
django-crispy-forms==2.6
django-extensions==4.1
django-tinymce==5.0.0
pillow==12.2.0
sqlparse==0.5.5
tzdata==2026.1
